+++ integration_tests/src/tests/mod.rs
@@ -2,7 +2,8 @@ mod authenticate_message_test;
 pub use authenticate_message_test::*;
 mod batch_onboarding;
 pub use batch_onboarding::*;
-// mod batch_onboarding_deals_test;
+mod batch_onboarding_deals_test;
+pub use batch_onboarding_deals_test::*;
 // mod change_beneficiary_test;
 // mod commit_post_test;
 // mod datacap_tests;
